refactor base classes for tests
