+++ libjava/lib/src/test/java/com/mapbox/services/geojson/PolygonTest.java
@@ -1,9 +1,16 @@
 package com.mapbox.services.geojson;
 
+import com.mapbox.services.commons.geojson.Polygon;
+import com.mapbox.services.commons.models.Position;
+
 import org.junit.Test;
+
+import java.io.IOException;
+import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
 
 /**
  * Created by antonio on 1/30/16.
@@ -12,32 +19,39 @@ public class PolygonTest extends BaseGeoJSON {
 
     @Test
     public void fromJson() {
-        com.mapbox.services.commons.geojson.Polygon geo = com.mapbox.services.commons.geojson.Polygon.fromJson(BaseGeoJSON.SAMPLE_POLYGON);
+        Polygon geo = Polygon.fromJson(BaseGeoJSON.SAMPLE_POLYGON);
         assertEquals(geo.getType(), "Polygon");
-        assertEquals(geo.getCoordinates().get(0).get(0).getLongitude(), 100.0, 0.0);
-        assertEquals(geo.getCoordinates().get(0).get(0).getLatitude(), 0.0, 0.0);
+        assertEquals(geo.getCoordinates().get(0).get(0).getLongitude(), 100.0, DELTA);
+        assertEquals(geo.getCoordinates().get(0).get(0).getLatitude(), 0.0, DELTA);
         assertFalse(geo.getCoordinates().get(0).get(0).hasAltitude());
     }
 
     @Test
     public void fromJsonHoles() {
-        com.mapbox.services.commons.geojson.Polygon geo = com.mapbox.services.commons.geojson.Polygon.fromJson(BaseGeoJSON.SAMPLE_POLYGON_HOLES);
+        Polygon geo = Polygon.fromJson(BaseGeoJSON.SAMPLE_POLYGON_HOLES);
         assertEquals(geo.getType(), "Polygon");
-        assertEquals(geo.getCoordinates().get(0).get(0).getLongitude(), 100.0, 0.0);
-        assertEquals(geo.getCoordinates().get(0).get(0).getLatitude(), 0.0, 0.0);
+        assertEquals(geo.getCoordinates().get(0).get(0).getLongitude(), 100.0, DELTA);
+        assertEquals(geo.getCoordinates().get(0).get(0).getLatitude(), 0.0, DELTA);
         assertFalse(geo.getCoordinates().get(0).get(0).hasAltitude());
     }
 
     @Test
     public void toJson() {
-        com.mapbox.services.commons.geojson.Polygon geo = com.mapbox.services.commons.geojson.Polygon.fromJson(BaseGeoJSON.SAMPLE_POLYGON);
+        Polygon geo = Polygon.fromJson(BaseGeoJSON.SAMPLE_POLYGON);
         compareJson(BaseGeoJSON.SAMPLE_POLYGON, geo.toJson());
     }
 
     @Test
     public void toJsonHoles() {
-        com.mapbox.services.commons.geojson.Polygon geo = com.mapbox.services.commons.geojson.Polygon.fromJson(BaseGeoJSON.SAMPLE_POLYGON_HOLES);
+        Polygon geo = Polygon.fromJson(BaseGeoJSON.SAMPLE_POLYGON_HOLES);
         compareJson(BaseGeoJSON.SAMPLE_POLYGON_HOLES, geo.toJson());
+    }
+
+    @Test
+    public void testTurfFixture() throws IOException {
+        Polygon polyHole = Polygon.fromJson(loadJsonFixture("turf-inside", "poly-with-hole.geojson"));
+        List<List<Position>> coordinates = polyHole.getCoordinates();
+        assertEquals(coordinates.size(), 1);
     }
 
 }
